move tests to separate pbts test file
